Auto_Proposal_WebAPI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
reportlab==4.0.7
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
pymysql==1.1.0
psycopg2-binary==2.9.9
python-jose==3.3.0
gspread==6.2.1
oauth2client==4.1.3
pandas==2.1.2
wfastcgi==3.0.0
asgiref==3.7.2
openpyxl==3.1.2
bcrypt==4.1.1
cryptography==41.0.7
cloud-sql-python-connector==1.5.0